Add support for flux
Add configure check X_AC_FLUX based on X_AC_LSF.

When the user does not specify a location, use pkg-config to determine
whether flux-core is installed and where.  Otherwise look for
flux/core.h and attempt to link against flux-core.so to use flux_open().

At runtime, look in the environment for FLUX_JOB_ID.

Determine whether to query the current flux instance or the parent
for the expiration time of the allocation.  Note that this
currently works by checking the environment for FLUX_KVS_NAMESPACE,
but flux will provide a more explicit mechanism in the future.
See https://github.com/flux-framework/flux-core/issues/3817
for details and status.

Fetch the expiration time and calculate remaining time based on that.

Support get_rank() by looking in the environment for FLUX_TASK_RANK.